# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a lead scrapping system API.",
        "title": "Lead Scrapper API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate user and return access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api_dto.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth_dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_dto.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api_dto.ApiResponse"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Create a new user account with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Signup Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_dto.SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_dto.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_dto.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_dto.ApiResponse"
                        }
                    }
                }
            }
        },
        "/lead/scrap": {
            "post": {
                "description": "Start a lead scrapping job for the specified industry and location across multiple sources (Google Maps, LinkedIn, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lead"
                ],
                "summary": "Scrap leads from multiple sources",
                "parameters": [
                    {
                        "description": "Scrap Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lead_dto.LeadScrapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_dto.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_dto.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_dto.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api_dto.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "auth_dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth_dto.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "auth_dto.SignupRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "lead_dto.LeadScrapRequest": {
            "type": "object",
            "required": [
                "industry_type",
                "location",
                "source"
            ],
            "properties": {
                "industry_type": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "source": {
                    "description": "NumberOfRequest int64        `json:\"number_of_request\" validate:\"required\"`",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lead_dto.SourceData"
                    }
                }
            }
        },
        "lead_dto.Source": {
            "type": "string",
            "enum": [
                "google_maps",
                "linked_in",
                "facebook",
                "instagram"
            ],
            "x-enum-varnames": [
                "GoogleMap",
                "LinkedIn",
                "Facebook",
                "Instagram"
            ]
        },
        "lead_dto.SourceData": {
            "type": "object",
            "required": [
                "number_of_request",
                "source"
            ],
            "properties": {
                "number_of_request": {
                    "type": "integer"
                },
                "source": {
                    "$ref": "#/definitions/lead_dto.Source"
                }
            }
        }
    }
}